requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
bibtexparser==1.4.3
blake3==1.0.5
build==1.3.0
cachetools==5.5.2
cbor2==5.6.5
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.15
ci-info==0.3.0
click==8.2.1
cloudpickle==3.1.1
coloredlogs==15.0.1
compressed-tensors==0.10.2
configobj==5.0.9
configparser==7.2.0
dataclasses-json==0.6.7
datasets==4.0.0
depyf==0.19.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
durationpy==0.10
einops==0.8.1
email_validator==2.2.0
et_xmlfile==2.0.0
etelemetry==0.3.1
eval_type_backport==0.2.2
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
gguf==0.17.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
interegular==0.3.3
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.26
langchain-chroma==0.2.5
langchain-community==0.3.27
langchain-core==0.3.72
langchain-mistralai==0.2.11
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langsmith==0.4.5
lark==1.2.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
looseversion==1.3.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mistral_common==1.8.3
mistralai==1.9.3
mlx==0.27.1
mlx-lm==0.26.2
mlx-metal==0.27.1
mmh3==5.2.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.5
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.10.0
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.90.0
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.10.18
outlines_core==0.2.10
overrides==7.7.0
packaging==24.2
pandas==2.3.1
partial-json-parser==0.2.1.1.post6
pathlib==1.0.1
pillow==11.3.0
posthog==5.4.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
propcache==0.3.2
protobuf==6.31.1
prov==2.1.1
psutil==7.0.0
puremagic==1.30
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.26.3
pymupdf4llm==0.0.27
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.0.1
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scipy==1.16.0
sentencepiece==0.2.0
sentry-sdk==2.34.1
setuptools==79.0.1
shapely==2.1.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.41
starlette==0.47.2
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.4
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
traits==7.0.2
transformers==4.54.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vllm==0.10.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
xgrammar==0.1.21
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
